+++ src/posts/posts.service.ts
@@ -1,11 +1,15 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { PostsRepository } from './posts.repository';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
+import { PublicationsRepository } from 'src/publications/publications.repository';
 
 @Injectable()
 export class PostsService {
-    constructor(private readonly postsRepository: PostsRepository) {}
+    constructor(
+        private readonly postsRepository: PostsRepository,
+        private readonly publicationRepository: PublicationsRepository 
+    ) {}
 
     async create(createPostDto: CreatePostDto) {
         return this.postsRepository.create(createPostDto);
@@ -32,6 +36,9 @@ export class PostsService {
     async delete(id: number){
         const post = await this.postsRepository.findOne(id);
         if(!post) throw new NotFoundException();
+
+        const publications = await this.publicationRepository.countPublicationsByPostId(id);
+        if(publications > 0) throw new ForbiddenException();
 
         return await this.postsRepository.delete(id);
     }
